[FIX] account: synchronized between analytic item and journal item
When modifying or deleting an Analytic Account on an Analytic Item, it
is not reflected on the Analytic Distribution of the Journal Item.
The user can be mislead, given on the business objects (invoice lines),
he sees the initial Analytic Distribution but there is no guarantee
that the related Analytic Item remains consistent.

task-4378407

closes odoo/odoo#199020

Signed-off-by: William André (wan) <wan@odoo.com>