[IMP] l10n_my_edi_extended: Malaysian fields in the portal
The POS supports to get your invoice using a QR code present
on a POS ticket after the fact.

Support for it is not easily done for public users and will
be supported later; but for logged in users we can display
the fields on their portal to allow them to fetch the invoice
correctly by themselves.

Part-of: odoo/odoo#215786
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>