[FIX] discuss: prevent race condition traceback in p2p
Before this commit, there could be a race condition where
the peer connection gets closed while an offer is being
negotiated.

This could happen for example when offers are being sent to
establish a p2p connection and the fallback to the SFU kicks in
and clears the connection.

This would lead to a traceback but no functional issue.

closes odoo/odoo#215995

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>