[FIX] project_timesheet_holidays: Allow timesheet recording for 1 day leaves
After PR 209570 , the way work intervals are computed might not reflect real working time for flexible employees, since the virtual schedule is based on a weekly span.

In this PR we chose to allocate a timesheet for single day leaves corresponding to average hours per day to be consistent with hr_holidays calculation.

closes odoo/odoo#218353

Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>