[FIX] marketing_card: translate datetime fields into the right timezone
It is generally desirable to display the date in the timezone of
relevant to the record when sending out cards to share.

Notably for event models, everyone will be participating relative to the
same timezone.

If the model where a field is fetched defines some timezone, we
translate it and show the datetime or date as it is in that timezone.
Rather than UTC.

task-4936298

Part-of: odoo/odoo#218633
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>