[FIX] account: add notify_progress to auto send CRON
The scheduled action for the automatic sending of invoices currently
does not explictly use the `_notify_progress` method introduced since
Odoo 18 to track the progress of CRON jobs.

This means that implicitly, by default the CRON will log
`processed 0 records, 0 records remaining` at each run.

This can be confusing, as the automatic invoice sending CRON is batched,
and might run multiple times in sucession, reporting each time that it
processed 0 records (which might not be true).

Proposed fix:
We correctly count the total account moves that need to be processed
(using a `search_count`) and calculate the number of processed records
and the remaining ones.

opw-4926541

closes odoo/odoo#218763

Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>