[FIX] account_edi_ubl_cii: Add SellerItemIdentification and add tests
Before the UBL refactor, the SellerItemIdentification node was populated
with the product code. We restore this behaviour.

In addition, the refactor added the StandardItemIdentification node
without specifying the schemeID (which is required by the schematron),
then commit f46c10f03e59e8 added `schemeID="0160"`.

This commit adds tests to enforce this behaviour.

task-none

Part-of: odoo/odoo#219924
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>